+++ src/common/check_migration_status/test.py
@@ -2,13 +2,15 @@ import subprocess
 from os import path
 import json
 
-input_path = "/openproblems-v2"
+input_sha = meta["resources_dir"] + "resources_test/common/task_metadata/input_git_sha.json"
+input_method_info = meta["resources_dir"] +  "resources_test/common/task_metadata/method_info.json"
 output_path = "output.json"
 
 cmd = [
     meta['executable'],
-    "--input", input_path,
-    "--output", output_path
+    "--git_sha", input_sha,
+    "--comp_info", input_method_info,
+    "--output", output_path,
 ]
 
 print(">> Running script as test")
@@ -20,6 +22,6 @@ assert path.exists(output_path)
 print(">> Reading json file")
 with open(output_path, 'r') as f:
     out = json.load(f)
-    print(out[0])
+    print(out)
 
 print("All checks succeeded!")